refactor: move assignment store to separate file
